add Brightness library. Fix a bag: need init after some time.

+++ arduino/arduino.ino
@@ -12,13 +12,14 @@ long temp = 0;
 float Vref = 5000;
 float Counts = 1024;
 float coeff = Vref / Counts;
+
 void loop()
 {
 	
 	holdingRegs[0] = 200;
 	
 
-	for (uint i = 0; i < 6; i++)
+	for (uint8_t i = 0; i < 6; i++)
 	{
 		temp = (analogRead(i)* coeff);
 		holdingRegs[i+1] = (uint16_t)temp;
